Upgrading to OpenLayers 3.20.1
Something is wrong with the way we’re using the 4.x branch that it
breaks in a fairly horrid way.

There’s nowt major we need in 4.x yet, so let’s ignore it.

If we need it then the other React-OpenLayers implementation might fix
it for us for free (see #79).