Switch MapUI to using the layer key (aka layerId) as it's unique key for React
We can't use hash because two layers in a map can have the same hash.

So this is the best bet short of creating immutable uuids when we make a layer.

P.S. This code will probably cause issues when we implement layer reordering because that'll change the layer's position in the array and, ergo, its key.